+++ src/nodes/edge_ai_pipeline/processors/pose_detection.py
@@ -1,20 +1,26 @@
 import copy
+import logging
 
-import dearpygui.dearpygui as dpg
-
+from gui.constants import Attribute, PinShape
 from models.mediapipe_hands.model import MediaPipeHands
 from models.mediapipe_pose.model import MediaPipePose
 from models.onnx_movenet_multipose_lightning.model import MoveNetMPL
 from nodes.edge_ai_pipeline.base import BaseNode
 
+try:
+    import dearpygui.dearpygui as dpg
+except ImportError:
+    pass
+
 
 class EdgeAINode(BaseNode):
-    def __init__(self, settings):
+    def __init__(self, settings, logger=logging.getLogger(__name__)):
         self.version = "0.1.0"
         self.name = "Pose Detection"
         self.theme_titlebar = [102, 51, 0]
         self.theme_titlebar_selected = [153, 76, 0]
         self.settings = settings
+        self.logger = logger
         self.configs = {}
         self.configs["models"] = {
             "MoveNet MulitPose Lightning": MoveNetMPL,
@@ -22,90 +28,96 @@ class EdgeAINode(BaseNode):
             "MediaPipe Pose": MediaPipePose,
         }
         self.configs["instances"] = {}
+        self.forms = {}
+        self.forms["model"] = {}
 
     def add_node(self, parent, node_id, pos):
         # Describe node attribute tags
         dpg_node_tag = str(node_id) + ":" + self.name.lower().replace(" ", "_")
         dpg_pin_tags = self.get_tag_list(dpg_node_tag)
+        self.forms["model"][dpg_node_tag] = list(self.configs["models"].keys())[0]
 
-        # Add a dynamic texture and a raw texture
-        with dpg.texture_registry(show=False):
-            dpg.add_raw_texture(
-                self.settings["node_width"],
-                self.settings["node_height"],
-                self.get_blank_texture(
-                    self.settings["node_width"], self.settings["node_height"]
-                ),
-                tag=dpg_node_tag + ":texture",
-                format=dpg.mvFormat_Float_rgba,
-            )
-
-        # Add a node to a node editor
-        with dpg.node(
-            tag=dpg_node_tag, parent=parent, label=self.name, pos=pos
-        ) as dpg_node:
-            # Set node color
-            with dpg.theme() as dpg_theme:
-                with dpg.theme_component(dpg.mvNode):
-                    dpg.add_theme_color(
-                        dpg.mvNodeCol_TitleBar,
-                        self.theme_titlebar,
-                        category=dpg.mvThemeCat_Nodes,
-                    )
-                    dpg.add_theme_color(
-                        dpg.mvNodeCol_TitleBarHovered,
-                        self.theme_titlebar_selected,
-                        category=dpg.mvThemeCat_Nodes,
-                    )
-                    dpg.add_theme_color(
-                        dpg.mvNodeCol_TitleBarSelected,
-                        self.theme_titlebar_selected,
-                        category=dpg.mvThemeCat_Nodes,
-                    )
-                    dpg.add_theme_color(
-                        dpg.mvNodeCol_NodeOutline,
-                        self.theme_titlebar,
-                        category=dpg.mvThemeCat_Nodes,
-                    )
-                    dpg.bind_item_theme(dpg_node, dpg_theme)
-
-            # Add pins that allows linking inputs and outputs
-            with dpg.node_attribute(
-                attribute_type=dpg.mvNode_Attr_Input,
-                tag=dpg_pin_tags[self.VIDEO_IN],
-            ):
-                dpg.add_text("VIDEO IN")
-            with dpg.node_attribute(
-                attribute_type=dpg.mvNode_Attr_Input,
-                shape=dpg.mvNode_PinShape_Quad,
-                tag=dpg_pin_tags[self.MESSAGE_IN],
-            ):
-                dpg.add_text("MESSAGE IN")
-            with dpg.node_attribute(
-                attribute_type=dpg.mvNode_Attr_Output,
-                tag=dpg_pin_tags[self.VIDEO_OUT],
-            ):
-                dpg.add_text("VIDEO OUT")
-            with dpg.node_attribute(
-                attribute_type=dpg.mvNode_Attr_Output,
-                shape=dpg.mvNode_PinShape_Quad,
-                tag=dpg_pin_tags[self.MESSAGE_OUT],
-            ):
-                dpg.add_text("MESSAGE OUT")
-
-            # Add a combo dropdown that allows selecting model
-            with dpg.node_attribute(attribute_type=dpg.mvNode_Attr_Static):
-                dpg.add_spacer(height=5)
-                dpg.add_combo(
-                    list(self.configs["models"].keys()),
-                    default_value=list(self.configs["models"].keys())[0],
-                    width=self.settings["node_width"],
-                    tag=dpg_node_tag + ":model",
+        if self.settings["gui"]:
+            # Add a dynamic texture and a raw texture
+            with dpg.texture_registry(show=False):
+                dpg.add_raw_texture(
+                    self.settings["node_width"],
+                    self.settings["node_height"],
+                    self.get_blank_texture(
+                        self.settings["node_width"], self.settings["node_height"]
+                    ),
+                    tag=dpg_node_tag + ":texture",
+                    format=dpg.mvFormat_Float_rgba,
                 )
 
-            # Add an image from a specified texture
-            with dpg.node_attribute(attribute_type=dpg.mvNode_Attr_Static):
-                dpg.add_image(dpg_node_tag + ":texture")
+            # Add a node to a node editor
+            with dpg.node(
+                tag=dpg_node_tag, parent=parent, label=self.name, pos=pos
+            ) as dpg_node:
+                # Set node color
+                with dpg.theme() as dpg_theme:
+                    with dpg.theme_component(dpg.mvNode):
+                        dpg.add_theme_color(
+                            dpg.mvNodeCol_TitleBar,
+                            self.theme_titlebar,
+                            category=dpg.mvThemeCat_Nodes,
+                        )
+                        dpg.add_theme_color(
+                            dpg.mvNodeCol_TitleBarHovered,
+                            self.theme_titlebar_selected,
+                            category=dpg.mvThemeCat_Nodes,
+                        )
+                        dpg.add_theme_color(
+                            dpg.mvNodeCol_TitleBarSelected,
+                            self.theme_titlebar_selected,
+                            category=dpg.mvThemeCat_Nodes,
+                        )
+                        dpg.add_theme_color(
+                            dpg.mvNodeCol_NodeOutline,
+                            self.theme_titlebar,
+                            category=dpg.mvThemeCat_Nodes,
+                        )
+                        dpg.bind_item_theme(dpg_node, dpg_theme)
+
+                # Add pins that allows linking inputs and outputs
+                with dpg.node_attribute(
+                    attribute_type=int(Attribute.INPUT),
+                    tag=dpg_pin_tags[self.VIDEO_IN],
+                ):
+                    dpg.add_text("VIDEO IN")
+                with dpg.node_attribute(
+                    attribute_type=int(Attribute.INPUT),
+                    shape=int(PinShape.QUAD),
+                    tag=dpg_pin_tags[self.MESSAGE_IN],
+                ):
+                    dpg.add_text("MESSAGE IN")
+                with dpg.node_attribute(
+                    attribute_type=int(Attribute.OUTPUT),
+                    tag=dpg_pin_tags[self.VIDEO_OUT],
+                ):
+                    dpg.add_text("VIDEO OUT")
+                with dpg.node_attribute(
+                    attribute_type=int(Attribute.OUTPUT),
+                    shape=int(PinShape.QUAD),
+                    tag=dpg_pin_tags[self.MESSAGE_OUT],
+                ):
+                    dpg.add_text("MESSAGE OUT")
+
+                # Add a combo dropdown that allows selecting model
+                with dpg.node_attribute(attribute_type=int(Attribute.STATIC)):
+                    dpg.add_spacer(height=5)
+                    dpg.add_combo(
+                        list(self.configs["models"].keys()),
+                        default_value=self.forms["model"][dpg_node_tag],
+                        width=self.settings["node_width"],
+                        callback=self.callback_change_model,
+                        user_data=dpg_node_tag,
+                        tag=dpg_node_tag + ":model",
+                    )
+
+                # Add an image from a specified texture
+                with dpg.node_attribute(attribute_type=int(Attribute.STATIC)):
+                    dpg.add_image(dpg_node_tag + ":texture")
 
         # Return Dear PyGui Tag
         return dpg_node_tag
@@ -114,25 +126,17 @@ class EdgeAINode(BaseNode):
         dpg_node_tag = str(node_id) + ":" + self.name.lower().replace(" ", "_")
         frame = None
         message = None
-        model_class = self.configs["models"][
-            dpg.get_value(dpg_node_tag + ":model")
-            if dpg.does_item_exist(dpg_node_tag + ":model")
-            else None
-        ]
-        model_name = (
-            dpg.get_value(dpg_node_tag + ":model")
-            if dpg.does_item_exist(dpg_node_tag + ":model")
-            else None
-        )
+        model_name = self.forms["model"][dpg_node_tag]
+        model_class = self.configs["models"][model_name]
 
         # Get linked node tag
         linked_node_tag = None
         for link in node_links:
             link_pin_shape = link[0].split(":")[2]
-            if link_pin_shape == str(dpg.mvNode_PinShape_Quad):
+            if link_pin_shape == str(int(PinShape.QUAD)):
                 linked_node_tag_message = ":".join(link[0].split(":")[:2])
                 message = node_messages.get(linked_node_tag_message, None)
-            if link_pin_shape == str(dpg.mvNode_PinShape_CircleFilled):
+            if link_pin_shape == str(int(PinShape.CIRCLE_FILLED)):
                 linked_node_tag = ":".join(link[0].split(":")[:2])
 
         # Inference
@@ -147,8 +151,9 @@ class EdgeAINode(BaseNode):
                 self.settings["node_width"],
                 self.settings["node_height"],
             )
-            if dpg.does_item_exist(dpg_node_tag + ":texture"):
-                dpg.set_value(dpg_node_tag + ":texture", texture)
+            if self.settings["gui"]:
+                if dpg.does_item_exist(dpg_node_tag + ":texture"):
+                    dpg.set_value(dpg_node_tag + ":texture", texture)
 
             # Generate message
             if message is None:
@@ -169,18 +174,26 @@ class EdgeAINode(BaseNode):
 
     def delete(self, node_id):
         dpg_node_tag = str(node_id) + ":" + self.name.lower().replace(" ", "_")
-        dpg.delete_item(dpg_node_tag + ":texture")
-        dpg.delete_item(dpg_node_tag)
+        if self.settings["gui"]:
+            dpg.delete_item(dpg_node_tag + ":texture")
+            dpg.delete_item(dpg_node_tag)
 
     def get_export_params(self, node_id):
         dpg_node_tag = str(node_id) + ":" + self.name.lower().replace(" ", "_")
         params = {}
-        params["version"] = self.version
-        params["position"] = dpg.get_item_pos(dpg_node_tag)
-        params["model"] = dpg.get_value(dpg_node_tag + ":model")
+        if self.settings["gui"]:
+            params["version"] = self.version
+            params["position"] = dpg.get_item_pos(dpg_node_tag)
+            params["model"] = self.forms["model"][dpg_node_tag]
         return params
 
     def set_import_params(self, node_id, params):
         dpg_node_tag = str(node_id) + ":" + self.name.lower().replace(" ", "_")
         if "model" in params:
-            dpg.set_value(dpg_node_tag + ":model", params["model"])
+            self.forms["model"][dpg_node_tag] = params["model"]
+            if self.settings["gui"]:
+                dpg.set_value(dpg_node_tag + ":model", params["model"])
+
+    def callback_change_model(self, sender, data, user_data):
+        dpg_node_tag = user_data
+        self.forms["model"][dpg_node_tag] = data
